fix-orders-component-and-misc-bugs
